+++ contracts/test/MockTerritories.sol
@@ -23,7 +23,7 @@ contract MockTerritories is ITerritories {
 
   function assignCombatants(
     uint256 clanId,
-    uint48[] calldata playerIds,
+    uint64[] calldata playerIds,
     uint256 combatantCooldownTimestamp,
     uint256 leaderPlayerId
   ) external override {}
